Move navigation/frontier operators from operators/core.py to environment/unknown_space/operators.py
These operators are specific to the frontier-based exploration domain and
only used by the frontier_search example and unknown_space tests.
