+++ packages/railroad/src/railroad/environment/unknown_space/__init__.py
@@ -1,6 +1,12 @@
 """UnknownSpaceEnvironment and helpers for frontier-based exploration."""
 
 from .environment import UnknownSpaceEnvironment
+from .operators import (
+    construct_move_navigable_operator,
+    construct_observe_site_operator,
+    construct_search_at_site_operator,
+    construct_search_frontier_operator,
+)
 from .types import Frontier, NavigationConfig
 from ..types import Pose
 
@@ -9,4 +15,8 @@ __all__ = [
     "NavigationConfig",
     "Pose",
     "UnknownSpaceEnvironment",
+    "construct_move_navigable_operator",
+    "construct_observe_site_operator",
+    "construct_search_at_site_operator",
+    "construct_search_frontier_operator",
 ]
